feat: change names and organise folders
