+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@solfacil/components-ui",
   "author": "Solfacil - Fernando Jesus",
-  "version": "0.1.43",
+  "version": "0.1.45",
   "private": false,
   "license": "MIT",
   "description": "Component library based on Vue and Tailwind",
