+++ src/stories/Introduction.stories.mdx
@@ -1,7 +1,5 @@
 import { Meta, Preview, Story } from '@storybook/addon-docs/blocks';
 
-<Meta
-  title="Getting Started/Installation"
-/>
+<Meta title="Getting Started/Installation" />
 
 # Installation
